// src/types/era/IEraAsset.ts
export interface IEraAsset {
    creator: string;
    quantity: number;
    item_type: string;
    isDivisible: boolean;
    description: string;
    scale: number;
    type1: number;
    type0: number;
    reference: string;
    item_type_sub: string;
    name: string;
    isConfirmed: boolean;
    key: number;
    timestamp: number;
}